Create 3Sum Problem
TimeComplexity O(N^2) and Space complexitty O(N)
Approch to solve above problem store the element of array in hash. 
loop 1 level nested loop and then check if sum of A[i] + A[j] is present in hash or not, if value is present in hash then sort the collection and store in set.
when you return then convert the hash into the arraylist.